+++ src/main/java/walkingkooka/spreadsheet/dominokit/condition/SpreadsheetConditionRightParserTokenComponentConditionFunction.java
@@ -17,8 +17,8 @@
 
 package walkingkooka.spreadsheet.dominokit.condition;
 
-import walkingkooka.spreadsheet.formula.ConditionRightSpreadsheetFormulaParserToken;
 import walkingkooka.spreadsheet.formula.SpreadsheetFormulaParsers;
+import walkingkooka.spreadsheet.formula.parser.ConditionRightSpreadsheetFormulaParserToken;
 import walkingkooka.spreadsheet.parser.SpreadsheetParserContext;
 
 import java.util.Objects;
