+++ src/main/java/walkingkooka/spreadsheet/dominokit/find/SpreadsheetFindDialogComponentSpreadsheetParserTokenVisitor.java
@@ -19,17 +19,17 @@ package walkingkooka.spreadsheet.dominokit.find;
 
 import walkingkooka.spreadsheet.dominokit.textmatch.TextMatchComponent;
 import walkingkooka.spreadsheet.expression.function.SpreadsheetExpressionFunctions;
-import walkingkooka.spreadsheet.formula.ConditionSpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.EqualsSpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.GreaterThanEqualsSpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.GreaterThanSpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.LessThanEqualsSpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.LessThanSpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.NamedFunctionSpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.NotEqualsSpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.SpreadsheetFormulaParserToken;
-import walkingkooka.spreadsheet.formula.SpreadsheetFormulaParserTokenVisitor;
-import walkingkooka.spreadsheet.formula.TextSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.ConditionSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.EqualsSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.GreaterThanEqualsSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.GreaterThanSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.LessThanEqualsSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.LessThanSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.NamedFunctionSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.NotEqualsSpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.SpreadsheetFormulaParserToken;
+import walkingkooka.spreadsheet.formula.parser.SpreadsheetFormulaParserTokenVisitor;
+import walkingkooka.spreadsheet.formula.parser.TextSpreadsheetFormulaParserToken;
 import walkingkooka.text.cursor.parser.ParserToken;
 import walkingkooka.tree.expression.ExpressionFunctionName;
 import walkingkooka.visit.Visiting;
